dspackagefolder/template/requirements.txt
toml
streamlit
pydantic
python-dotenv
